requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
fonttools==4.47.2
google-search-results==2.4.2
idna==3.6
kiwisolver==1.4.5
matplotlib==3.8.2
mkl-service==2.4.0
packaging==23.2
pillow==10.2.0
pyparsing==3.1.1
python-dotenv==1.0.0
requests==2.31.0
serpapi==0.1.5
setuptools==68.2.2
urllib3==2.1.0
wheel==0.41.2
XlsxWriter==3.1.9
